Add jobs command.
Required adding a new struct Shell that contains state info about the current instance of the shell. Currently this only holds the list of running jobs. I'm undecided as to whether or not it
should contain information such as the current working directory and environment variables. A benefit of containing that information would be the ability to hotswap the shell with a command.